requirements.txt
# The Voice Codex — pinned dependencies
# Python >= 3.10 required
# Install: pip install -r requirements.txt

# ── Core audio / DSP ──────────────────────────────────────────────────────────
librosa==0.11.0
numpy==1.26.4
scipy==1.16.3
praat-parselmouth==0.4.7

# ── Data handling ─────────────────────────────────────────────────────────────
pandas==2.2.3

# ── Machine learning / analysis ───────────────────────────────────────────────
scikit-learn==1.8.0

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib==3.10.8
seaborn==0.13.2

# ── Notebook ─────────────────────────────────────────────────────────────────
jupyter==1.1.1
ipykernel==6.29.5

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==9.0.1
